"Refactor playlist and track services to use TrackInfo"
The PlaylistService.getTracksAvailable and TrackService.findTracksAvailableForPlaylist methods were refactored to use the TrackInfo object instead of the Track object. This change will allow additional track information to be presented without modifying the base Track object. Changes were also made in PlaylistController and playlist_info.html to accommodate for this shift in data type.

Additionally, minor UI enhancements were made to playlist_info.html, including the addition of Bootstrap Icons for the delete and add actions.

Also, the TrackInfo.getFullTitle formula was updated to handle the cases when the track composer might be null.